Crate wasm_spec_interpreter

This library provides a way to interpret Wasm functions in the official Wasm specification interpreter, written in OCaml, from Rust.

In order to not break Wasmtime’s build, this library will always compile. It does depend on certain tools (see README.md) that may or may not be available in the environment:

when the tools are available, we build and link to an OCaml static library (see with_library module)

when the tools are not available, this library will panic at runtime (see without_library module).

Structs

SpecInstance

Represents a WebAssembly instance from the OCaml interpreter side.

Enums

SpecExport

Represents a WebAssembly export from the OCaml interpreter side.

SpecValue

Enumerate the kinds of Wasm values the OCaml interpreter can handle.

Functions

export

instantiate

interpret

interpret_legacy

setup_ocaml_runtime

support_compiled_in

Check if the OCaml spec interpreter bindings will work.